Update get method to support parameters
Given that this SDK already supports parameters in POST and
DELETE requests, it seems, the consistent action will be to include
parameters in the GET request.

The only caveat here (which also applies for POST and DELETE) being
when the access token is explicitly supplied, the params array must
also be supplied (even as an empty array)

e.g. `$response = $fb->delete('/{post-id}', '{explicit_access_token}');`
The above will fail asking for the array since the second input
is the params array. So for the above to work it's actually

e.g. `$response = $fb->delete('/{post-id}', [], '{explicit_access_token}');`

Where [] is an empty array or can be supplied params such as user when
unblocking a page as mentioned in #379

The docs have been updated to mention this.